refactor EnergyLabeler Moved all functionality to the underlying objects of LandingZone and SecurityHub respectively.
